Break apart components
