+++ src/app/item-popup/SocketDetails.m.scss
@@ -38,56 +38,3 @@
   height: 16px;
   width: 16px;
 }
-
-.material {
-  composes: flexRow from '../dim-ui/common.m.scss';
-  align-items: center;
-  margin-top: 4px;
-  font-size: 10px;
-  img {
-    height: 16px;
-    width: 16px;
-  }
-}
-
-.selectedPlug {
-  composes: flexRow from '../dim-ui/common.m.scss';
-  display: grid;
-  grid-template-columns: var(--item-size) 1fr 1fr;
-  grid-template-areas:
-    'mod desc stat'
-    'mod modstat stat';
-  gap: 0 8px;
-  @include phone-portrait {
-    grid-template-columns: var(--item-size) 1fr;
-    grid-template-areas:
-      'mod desc'
-      'mod modstat'
-      'mod stat';
-    gap: 8px 8px;
-  }
-}
-
-.modIcon {
-  composes: flexColumn from '../dim-ui/common.m.scss';
-  align-items: center;
-  grid-area: mod;
-}
-
-.modDescription {
-  composes: flexColumn from '../dim-ui/common.m.scss';
-  grid-area: desc;
-  h3 {
-    font-weight: bold;
-    margin: 0;
-  }
-}
-
-.modStats {
-  grid-area: modstat;
-}
-
-.itemStats {
-  grid-area: stat;
-  margin: 0;
-}
