+++ src/app/item-popup/SocketDetails.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { DimSocket, D2Item, DimPlug, DimStat } from 'app/inventory/item-types';
+import { DimSocket, D2Item } from 'app/inventory/item-types';
 import Sheet from 'app/dim-ui/Sheet';
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import {
@@ -22,9 +22,7 @@ import { createSelector } from 'reselect';
 import { itemsForPlugSet } from 'app/collections/PresentationNodeRoot';
 import idx from 'idx';
 import _ from 'lodash';
-import { interpolateStatValue } from 'app/inventory/store/stats';
-import { StatValue } from './PlugTooltip';
-import ItemStats from './ItemStats';
+import SocketDetailsSelectedPlug from './SocketDetailsSelectedPlug';
 
 interface ProvidedProps {
   item: D2Item;
@@ -179,7 +177,12 @@ function SocketDetails({ defs, item, socket, stores, unlockedPlugs, onClose }: P
   // TODO: maybe show them like the perk browser, as a tile with names!
 
   const footer = selectedPlug && (
-    <SelectedPlug plug={selectedPlug} defs={defs} item={item} currentPlug={socket.plug} />
+    <SocketDetailsSelectedPlug
+      plug={selectedPlug}
+      defs={defs}
+      item={item}
+      currentPlug={socket.plug}
+    />
   );
 
   return (
@@ -191,7 +194,7 @@ function SocketDetails({ defs, item, socket, stores, unlockedPlugs, onClose }: P
     >
       <div className={clsx('sub-bucket', styles.modList)}>
         {mods.map((mod) => (
-          <Mod
+          <SocketDetailsMod
             key={mod.hash}
             className={clsx({
               [styles.selected]: selectedPlug === mod,
@@ -210,8 +213,8 @@ function SocketDetails({ defs, item, socket, stores, unlockedPlugs, onClose }: P
 
 export default connect<StoreProps>(mapStateToProps)(SocketDetails);
 
-// TODO: use SVG!
-function Mod({
+// TODO: use SVG! make a common icon component!
+export function SocketDetailsMod({
   itemDef,
   defs,
   className,
@@ -243,109 +246,6 @@ function Mod({
           <div className="energyCost">{itemDef.plug.energyCost.energyCost}</div>
         </>
       )}
-    </div>
-  );
-}
-
-function SelectedPlug({
-  plug,
-  defs,
-  item,
-  currentPlug
-}: {
-  plug: DestinyInventoryItemDefinition;
-  defs: D2ManifestDefinitions;
-  item: D2Item;
-  currentPlug: DimPlug | null;
-}) {
-  const selectedPlugPerk =
-    (plug.perks && plug.perks.length > 0 && defs.SandboxPerk.get(plug.perks[0].perkHash)) ||
-    undefined;
-
-  const costStatHashes = [3578062600, 2399985800, 3344745325, 3779394102];
-
-  const materialRequirementSet =
-    (plug.plug.insertionMaterialRequirementHash &&
-      defs.MaterialRequirementSet.get(plug.plug.insertionMaterialRequirementHash)) ||
-    undefined;
-
-  const stats = _.compact(
-    plug.investmentStats.map((stat) => {
-      if (costStatHashes.includes(stat.statTypeHash)) {
-        return null;
-      }
-      const itemStat = item.stats && item.stats.find((s) => s.statHash === stat.statTypeHash);
-      if (!itemStat) {
-        return null;
-      }
-      // const statDef = defs.Stat.get(stat.statTypeHash);
-      const statGroupDef = defs.StatGroup.get(
-        defs.InventoryItem.get(item.hash).stats.statGroupHash!
-      );
-
-      const statDisplay =
-        statGroupDef && statGroupDef.scaledStats.find((s) => s.statHash === stat.statTypeHash);
-
-      const currentModValue =
-        (currentPlug && currentPlug.stats && currentPlug.stats[stat.statTypeHash]) || 0;
-
-      const updatedInvestmentValue = itemStat.investmentValue + stat.value - currentModValue;
-      let itemStatValue = updatedInvestmentValue;
-      let modValue = stat.value;
-      if (statDisplay) {
-        itemStatValue = interpolateStatValue(updatedInvestmentValue, statDisplay);
-        modValue =
-          itemStatValue - interpolateStatValue(updatedInvestmentValue - stat.value, statDisplay);
-      }
-
-      return {
-        modValue,
-        dimStat: {
-          ...itemStat,
-          value: itemStatValue
-        } as DimStat
-      };
-    })
-  );
-
-  return (
-    <div className={styles.selectedPlug}>
-      <div className={styles.modIcon}>
-        <Mod itemDef={plug} defs={defs} />
-        {materialRequirementSet &&
-          materialRequirementSet.materials.map((material) => {
-            const materialDef = defs.InventoryItem.get(material.itemHash);
-            return (
-              materialDef &&
-              material.count > 0 &&
-              !material.omitFromRequirements && (
-                <div className={styles.material} key={material.itemHash}>
-                  {material.count.toLocaleString()}
-                  <BungieImage
-                    src={materialDef.displayProperties.icon}
-                    title={materialDef.displayProperties.name}
-                  />
-                </div>
-              )
-            );
-          })}
-      </div>
-      <div className={styles.modDescription}>
-        <h3>{plug.displayProperties.name}</h3>
-        {selectedPlugPerk ? (
-          <div>{selectedPlugPerk.displayProperties.description}</div>
-        ) : (
-          plug.displayProperties.description && <div>{plug.displayProperties.description}</div>
-        )}
-      </div>
-      <div className={styles.modStats}>
-        {stats.map((stat) => (
-          <div className="plug-stats" key={stat.dimStat.statHash}>
-            <StatValue value={stat.modValue} defs={defs} statHash={stat.dimStat.statHash} />
-          </div>
-        ))}
-      </div>
-      <ItemStats stats={stats.map((s) => s.dimStat)} className={styles.itemStats} />
     </div>
   );
 }
