+++ src/app/item-popup/SocketDetailsSelectedPlug.m.scss.d.ts
@@ -1,12 +1,12 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'categoryIcon': string;
-  'energyElement': string;
-  'modList': string;
-  'notUnlocked': string;
-  'selected': string;
-  'socketDetailsSheet': string;
+  'itemStats': string;
+  'material': string;
+  'modDescription': string;
+  'modIcon': string;
+  'modStats': string;
+  'selectedPlug': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
